Add a helper for recording file creations in a log entry
This commit adds a helper for logging file creations in a log entry.
The file from the source path is hard linked in to the log entry,
and a subsequent operation is added to hard link the file in the
storage from the log entry.

The existing file creation operations in the other helpers were
replaced to use this one.