Add a helper for recording flushes in a log entry
This commit exposes a helper method for adding a flush to a log entry.
Some of the more complex helpers add flushes themselves as they generally
map directly to what we'd do anyway when loggin a particular transaction.
The explicit flushing method will be used for composing transaction from
smaller primitives like RecordFileCreation.