Apply logged packs via physical logging protocol
The possible pack included with a transaction is currently handled
logically. The base name of the logged pack is included in the
transaction and is separately applied to the repository by linking
the files derived from the base name to the repository. This is
redundant handling as this could be handled via generic file system
operations that we log in the physical logging protocol.

This commit logs the file system operations to be performed in order
to apply the pack. As the pack prefix is no longer needed in the log
entry, it is removed.