Simplify transaction API for updating alternate
Transaction currently provides a method to stage an alternate update
into the transaction. The path is the relative path of the repository
that we want to link the alternate to.

This commit simplifies the external API to just marking that the file
was updated. This signals to the transaction manager to check the file
and to stage any updates.

This simplified API makes it easier to move the alternate updates to
be logged through the physical logging protocol by simply logging the
staged alternate file of the transaction.