Log physical operations to unlink an alternate
Alternate unlinking is currently logged as a logical operation. On
applying the log entry, all loose objects and the packs from the
alternate are linked to the repository, and the repository's
'objects/info/alternates' file is removed.

This commit removes the logical operation, and replaces it by logging
the physical operations that perform the same operation. Logging the
file system operations is a more composable approach.