git: Introduce `ReferenceBackend` types
The repos in Gitaly use files reference backend, as we introduce the new
reftable backend, it is important to identify which format the
repository is using.

Introduce the `ReferenceBackend` type which denotes the reference
backend used in the repository. This allows us to add backend specific
information. Also add code for identifying the backend type for a
provided repository.

Most of this is similar to how the `git.ObjectHash` mechanism works.