updateref: Use reference backend specific error regexes
The error regexes are hardcoded as per the files reference backend. As
we're introducing the reftables reference backend, it makes sense to
split the error regexes to be more backend specific.

Let's move the regexes to the `git.ReferenceBackend` structure and use
this inside the `updateref` package. Note that we currently copy the
same regexes as the files backend, which we will fix in the followin
commits.