Allow repo replication without alternates as a seed
By default, we seed a replicated repository with a snapshot that
includes alternates/pool data. However, if
ReplicateObjectDeduplicationNetworkMembership is set to true, we should
be able to skip the alternates from the snapshot as we would
replicate/sync the pool to the target storage and connect the replicated
repo to it.

This would be beneficial as it would speed up the replication process
and save up disk space that would otherwise have to be cleaned up (see
https://gitlab.com/gitlab-org/gitlab/-/issues/440443 for more
information).