git: Fix `RefLockedRegex` for reftable
The `RefLockedRegex` regex for reftable is incorrect, fix the regex and
re-enable the disabled tests. In the files backend this error was also
returning the reference which failed to obtain a lock. But with
reftables locks are on a table level and we do not receive any
information about which references failed to obtain a lock. Modify our
code accordingly.