git: Remove `PackedRefsLockedRegex` for reftable
The `PackedRefsLockedRegex` regex for reftable is not used, since
reftable uses the same error as the `RefLockedRegex` when the reftable
is locked.

Rename `TestUpdater_packedRefsLocked` to `TestUpdater_packRefsLocked`
since this is testing the running 'git-pack-refs(1)' and is not specific
to files backend related 'packed-refs'. Fix the test to also work with
reftables.