Write reference updates to disk in transaction manager tests
TransactionManager's tests are currently not writing the performed
reference updates to the disk and are rather just directly staging
them in the transaction. This commit changes the references to be
actually written to the disk. This becomes necessary as we're about
to log the physical operations to perform the reference changes of
a transaction.