Exercise more realistic scenarios in TransactionManager tests
TestTransactionManager/delete_reference_with_incorrect_old_tip is
testing that a reference won't be deleted if the old tip is incorrect.
Right now it's directly staging the incorrect reference update. Such
an update would not be staged in practice as the Git command the
transaction would run against its snapshot would also fail due to the
reference's old value being wrong already in the snapshot.

This commit changes the test exercise a more realistic scenario where
a concurrent transaction changed the reference's value while the
reference deleting transaction is staging its changes.

Similar issue is present in multiple tests and we amend them all to
exercise concurrency properly instead of directly staging incorrect
updates.

This changes enables us to later write the reference updates on the
disk in the tests as well. Currently they're just staged in-memory.
In practice, they'd always be ran through a Git command that performs
the verification already in the snapshot as the reference updates
are captured from the reference-transaction hook.