Remove unrealistic TransactionManager tests
This commit removes a few tests that are exercising scenarios that would
not happen in reality.

- Tests creating a file-directory conflict within the same transaction are removed.
  The Git command the transaction is running against the snapshot would already
  fail as it's not legal to create such a reference transaction.
- Test checking that branches can't point to non-commit objects is removed.
  The Git command the transaction is running would again fail when attempting to
  do so. We wouldn't attempt to commit such a transaction.
- Test creating references to missing objects are removed. The transaction would
  fail already as the Git command updating the references would again fail due
  to the missing objects.

All in all, we can expect that each transaction is correct in isolation. The
TransactionManager should only focus on verifying that concurrenty transactions
are not making conflicting changes. As such, tests and verification logic related
to invalid state changes are not necessary in TransactionManager.