housekeeping: Reword skip in `TestRepositoryManager_CleanStaleData_referenceLocks`
The test `TestRepositoryManager_CleanStaleData_referenceLocks` is
skipped with reftables because this is files backend specific.

The reftable backend is planned to rollout after the transaction
manager. Since the transaction manager will control how compaction is
run on reftables, there is no scenario where stale locks will be left
over. So we don't have to clean up stale locks similar to how it's done
in files backend.