Test recovering from write-ahead log by panicking
TestTransactionManager/recovers_from_write_ahead_log_on_start_up
tests that the manager recovers a write from the write ahead log
on start up. It achieves this by shutting down the manager in the
beforeApplyLogEntry hook to see if the manager applies the log entry
properly after restarting. Shutting down the manager cancels the
execution context which leads to the 'update-ref' command used to
apply the reference updates to terminate and bubble up the error.
This will no longer work soon as we're about to log the reference
updates as the file system operations to perform and don't involve
an update-ref process in the application process. This commit changes
the test to panic in the hook which triggers immediate shutdown of the
TransactionManager. The effect is the same but it doesn't depend on
the context cancellation being respected in a particular location.

Similar problem exists in two other test cases so we apply the same
fixes to those test cases as well.