test: Align SpawnToken timeout with default value
The `TestGetSpawnToken_CommandStats` test was marked as flaky several
weeks back: https://gitlab.com/gitlab-org/gitaly/-/issues/5791

The flake is apparently happening due to resource contention within the
CI runner, and not a fault of the test logic itself. Investigation of
production metrics showed that `GetSpawnToken` may time out during
periods of high disk I/O, but it is not a common issue.

Increase the timeout specified in the test to match the default value of
10 seconds. Hopefully this eliminates any flake.