Simplify transaction API for default branch updates
Transaction currently provides a method to stage a default branch
update into the transaction. It takes in the actual value of to set
for the default branch.

This commit simplifies the external API to just marking that the default
branch was updated. This signals to the transaction manager to check the
file and to stage any updates.

This simplfied API is a better fit as the default branch updates are now
using the physical logging protocol. The transaction simply stages the
updated file in the repository, and the transaction manager picks includes
it in the committed log entry.