Implement a helper for building a reference tree
In order to figure out the changes each reference transaction has
done to the repository's reference state on the disk, we should
inspect the file system state after the reference transaction has
been applied. This commit implements a helper type that builds a
tree representation of references. The tree representation will in
a later commit help us walk reference state in correct order and
avoid walking the same directories twice. This makes it easy to
record the file system operations the reference transaction performed.