Extract commit-graphs rewriting support in housekeeping manager
We added the support for repacking housekeeping tasks. After running
repacking, the commit-graphs should be updated. So, in the MR that adds
the support for housekeeping, the commit-graphs are updated as a part
of the repacking housekeeping task. However, housekeeping manager has
various heuristics to determine if a repository's commit-graphs should
be rewritten. The previous work only covers a portion of them, quite
close but it doesn't match 100%. For example, it ignores previous
commit-graph conditions such as disabled bloom filters. The heuristics
of housekeeping manager can be found here [1].

The housekeeping in WAL must be fully compatible with the traditional
housekeeping manager. We need to add a new housekeeping task to update
the commit-graph. The scheduler should be the one who decides if the
commit-graph need to be updated. The manager can get rid of that logic
and focus on the execution only.

This commit extracts commit-graph rewriting support from the Repacking
task to its own task.

[1] https://gitlab.com/gitlab-org/gitaly/blob/423efd6731985d75009ac4692f36ad039334eb01/internal/git/housekeeping/optimization_strategy.go#L234