stat: Add `ReferenceBackend` to the repository stats struct
Add a `ReferenceBackend` field to the `ReferencesInfo` struct, this will
be used by the housekeeping strategy to consider if we want to create
commit graphs or not.

For the files backend, we check if there are loose refs or packed-refs
present before writing commit graphs. This logic doesn't carry over to
the reftable backend. So let's always write the commit graph for the
reftable backend.

Fix all the tests around this.