repository: Use `NOT_FOUND` gRPC code with `WriteRef` RPC
The `WriteRef` RPC creates or updates a reference in a repository. If
the requested new or old revision objects do not exist in the
repository, the RPC returns an `INTERNAL` gRPC error code.

Update the `WriteRef` RPC error handling to instead return a `NOT_FOUND`
gRPC error code with the `ReferenceNotFoundError` detail when the
request contains objects revisions that do not exist in the repository.