Remove reference validation name validation from TransactionManager
TransactionManager is currently validating the reference names the
transaction are attempting to update. This is unnecessary as we can
expect the transaction to stage changes only for references with
valid names. If they attempted to update references with invalid
names, the Git command they use to do so should already fail. Further,
we can assume the transactions are doing the correct thing in isolation
and we should only focus on conflict checking, not data validation.

This commit removes the reference name validation from TransacitonManager
as it is redundant. All reference should already have been validated
when the transaction is about to commit. We keep ensuring that all
references are in the 'refs' directory. 'git update-ref' would happily
accept such reference updates. While we shouldn't be accepting such
updates in general, we add check for this as a sanity check to ensure
our expectations are not being violated.