Remove lock clearing functionality in TransactionManager
This commit removes the lock clearing functionality of
TransactionManager. This was previously needed applying
reference updates could be interrupted which would leave
stale locks in the repository. We're soon moving to logging
the file system operations to update the reference. Doing
so, the references are directly updated without writing out
lock files first. After that, there's no longer a chance
of stale locks being left in the repository under any
circumstances.

Repositories could still have stale locks in them from time
before transactions. The lock clearing functionality that was
removed would not help with this as the transaction's snapshot
would also contain the locks and fail to update the references.
Existing stale locks in the repository will be handled through
a migration that is ran when a repository is first accessed with
transactions enabled.