Test pointing to a non-existent object with a more realistic scenario
TestTransactionManager/create_a_reference_to_non-existent_object is
testing that the transaction manager rejects references being pointed
to non-existent objects. Currently this is done by staging directly a
reference update to a non-existent object. This is not a realistic
scenario as Git would fail the update when the transaction is performing
it against its snapshot repository. This commit changes the scenario
to be tested more realistically by a concurrent pruning operation being
committed while the transaction is performing its reference update. This
is a scenario that could actually happen.

This enables us to later write the reference changes on the disk in the
tests as well as we're no longer staging an update that is invalid to
begin with.