Flush modifications performed when applying operations
The logged operations currently include flushes, and the logged flushes
are performed when applying a transaction's operations. While this is nice
in the sense that all operations are logged, it's not entirely necessary
to log the flushes as the node applying the operation can determine from
the operations which paths need flushing. When WAL archival comes around,
we might be loading large amounts of log entries from the archive. Applying
them is more efficient if we don't flush in between. In this case, logging
the flushes is pointless.

This commit adds logic in applyOperations to automatically flush all
dirtied paths.