Add a method for recording directory entry removals
This commit adds a method to the WAL entry helper that records a
removal of a directory entry. The call sites that were doing the
same operation were updated to use the new method as well.

The new method will be used initially in tests only and later in
production code as well.