Deduplicate objects when asserting repository contents
TransactionManager's tests assert contents of packfiles. In addition
to asserting the exact contents of the packfiles, it also asserts
that Git reads the same set of objects from the repository. This is
not currently accounting for objects that may be duplicated in the
packs. Git would not read out the same objects twice when printing
out the object set of the repository.

Deduplicate the objects in the pack before comparing them to the
set of objects seen by Git.