Enable transactions in FetchRemote
TestFetchRemote_pooledRepository is still being skipped with
transactions due to them historically not supporting object pools.
Object pools are now supported, so let's start running the test
as usual.

The test setup is not building the 'gitaly-hooks' binary which leads
to the reference changes not being recorded in the transaction.
Configure 'gitaly-hooks' to be correctly built in the test setup.

TestFetchRemote was skipped as it previously was not using the
reference transaction hook. Nowadays it is, so also enable the tests
for it.