repository: Make `GetSnapshot` backup reftables
The `GetSnapshot` RPC is used to backup objects, refs and other important
aspects of a git repository. In this sense it also makes sense to backup
reftabes as they're introduced.

Let's add the reftable folder to the backup directives and fix the tests
to consider reftables too. With this we can also remove the skip we
added to the `GetSnapshotTest`.