Enable TestObjectsSize test with a pool with transactions
This commit enables the TestObjectsSize subtest that is exercising
pools with transactions. Pools were previously not supported but
are now.

The test was assuming storage details by figuring out the expected
size of the objects from the disk expecting them to be loose objects.
As the TransactionManager packs the objects, their size will be
smaller. Adjust the test to account for this.