range-diff: support RawRangeDiff RPC
Since we want to be able to compare multiple push versions
of a MergeRequest in GitLab, we implement RawRangeDiffRPC
in Gitaly based on git-range-diff, which can be used to
output raw data for differences between two commit ranges.
It supports three different input parameters: RangePair,
RevisionRange, and BaseWithRevisions. The details of these
three input parameters are described in the description
section of https://git-scm.com/docs/git-range-diff#_description.

Signed-off-by: ZheNing Hu <adlternative@gmail.com>