Move (*housekeeping.RepositoryManager).CleanStaleData to another file
The mentioned function is located in
"internal/git/housekeeping/clean_stale_data.go" file. It might not be a
good location for it. Other functions doing similar things are located
in "internal/git/housekeeping/optimization_strategy.go". This file will
be extracted to a sub-package in some later commits. Thus, it makes
sense to move the mentioned function to that file so that all functions
of the manager can be moved as a whole later.