Add WAL transaction to the housekeeping manager
This commit integrates WAL transactioning system with the housekeeping
manager. For every RPC request reaching Gitaly, a middleware initiates a
transaction and injects it into the context of the request. However, For
`OptimizeRepository` RPC, which is a MAINTENANCE op, the middleware
doesn't initializes the transaction. The housekeeping manager initiates
the transaction and commits that transaction itself.

Apart from that, Gitaly starts a cron-like housekeeping scheduler. This
scheduler walks through the list of repositories and triggers
housekeeping randomly. Thus, this commit also adds transaction
initialization to the aforementioned scheduler.

The housekeeping manager skips some housekeeping tasks. Those tasks are
deemed unnecessary after the transaction is enabled. Two significant
tasks are:
- Pruning objects. The traditional housekeeping manager prunes all
objects exceeding a grace period. This period is to ensure an object is
not removed while it's used by another request. WAL handles parallel
requests in a very different way and unreachable objects are not
included. Eventually, they'll be removed by a full repack. Thus, no need
to call pruning independently.
- Clean stale data. A WAL transaction copies meaningful data over when
it is applied to the destination repository. All temporary files, dirs,
locks, etc. are removed. Thus, there should be no stale data.