gitalybackup: Move restore specific client methods into restore command
The backup type Manager has an abstraction around repositories in order
to swap out between a local implementation and an RPC implementation.
When the restore repository cleanup code was added, this abstraction was
broken by adding RPC only methods to backup.Manager.

This change begins fixing this by hoisting the RPC calls back into the
restore command. So that backup.Manager can be cleaned up.