Generate transaction's pack in its staging directory
We're currently generating the transaction's pack directly in the
WAL entry's directory we are staging. This would lead to some unneeded
directory entries being logged. Generate the files into the root of
the transaction's staging directory instead so they are not part of
the log entry before being explicitly added to it.