backup: Add LogEntryArchiver for WAL entries
One of the key benefits of creating a write ahead log for git operations
is that it allows us to replay log entries as a granular backup.
Currently the `TransactionManager` will delete entries once they are no
longer needed, we want to be able to archive them prior to cleanup. The
new `LogConsumer` and `LogManager` interfaces added in the previous
commit provide a way for external packages to hook into the WAL and
operate against committed log entries.

Add a new `LogEntryArchiver` which will be notified of newly available
entries, `tar` the entry directory, and write it to a `backup.Sink`,
which may be either a local directory or object storage. It has a
configurable number of worker goroutines to perform the backups.

A given partition may only have one entry being backed up at a time, and
entries are always processed in order. The `LogManager` will provide the
range of all LSNs that are available for backup. On startup this may
include LSNs that were backed up by a previous archiver. To avoid
duplicating that work, we check if the there is an existing blob / file
for the LSN before proceeding to create an archive.

Should a successfully backed up entry be resubmitted to the
`LogEntryArchiver`, for example if the `TransactionManager` shutdown
before reading the `AcknowledgeTransaction` response, the archiver will
acknowledge the transaction a second time without creating a new backup.

Failures will typically share a common cause, such as connectivity
issues with object storage or overtaxed system resources. The archiver
will wait with an exponential backoff for subsequent failures before
continuing on.

Entries are stored by <PARTITION_ID>/<LSN>, which allows access to the
entries of a given repository or fork cluster based on their
partition ID.

Partition IDs for a given path are not currently stable between Gitaly
nodes, so migrating a repository to a new node will result in a new
partition ID that does not match the existing backups. This also implies
that Gitaly nodes cannot share the same object storage bucket for WAL
backups as their partition IDs will overlap and will overwrite each
others backups.