Move housekeeping configs to a new package
Recently, we added support for running housekeeping tasks in the
context of write-ahead logging. This work uses some functionalities of
the housekeeping package directly. At recent, those additions have been
isolated in the transaction manager. We are looking forward to
integrating the housekeeping-in-transaction with the existing
housekeeping manager. We'll use storagectx to extract the transaction
from the current context.

Unfortunately, the housekeeping package uses some internal git-related
packages. Those packages use storagectx package to extract the
transaction from the context. So, we cannot add housekeeping-related
functions to storagectx. Otherwise, a circle import is formed. This
commit moves the housekeeping config out of the housekeeping package to
its own package.