Write full repack timestamp file after applying a transaction
The housekeeping manager depends on a timestamp file to determine if it
needs to run a full repacking task. After running such task, the
timestamp is updated accordingly. However, the transaction manager
doesn't update the timestamp and will always trigger a full repack task.
This commit writes that file after a full repack task is applied in the
transaction manager.