Add a helper for recording reference changes
We're moving to a physical logging protocol that record the file
system operations that need to be performed in order to apply a
tranactions. Reference changes are one of the remaining operations
to migrate

This commit adds a helper for recording the file system operations
to apply reference changes. The file system operations to perform
are recorded by first recording the pre-image state, then applying
the reference changes, and then finally recording the post-image
state. Recording the changes like this ensures we're setting up the
directories and files as Git would.