Sync log entry completely in the critical section
TransactionManager has so far been flushing the log entry to the disk
before queuing it for commit. As we are now staging some file only after
the critical section, such as with references, we should sync the staged
files to disk before committing the log entry. This is unideal as this
means we are writing files to disk in the critical section which is bad
for performance. The file writes are sequential instead of parallel. This
is due to lacking proper conflict checks that we could use to ensure that
all changes staged by transactions do not conflict with concurrently
committed transactions.

Until we have the conflict checks in place, sync the log entry completely
before logging it to ensure the state staged during the critical section
is also flushed.