Fix flaky test TestPipeline/context_done
TestPipeline/context_done is testing that the context cancellations
are respected when waiting for a job to be picked up by a worker.
Unfortunately the test is flaky as the select statement where we are
waiting for a job to be picked up or the context being canceled is
indeterministic. This commit fixes the issue by overriding the channel
the job is being sent to with a channel where it will never be received
from.