Fix flaky test TestInfoRefsUploadPack_cache
TestInfoRefsUploadPack_cache is flaky as it creates the invalid
repository in a shared directory that other invocations of the same
tests with different feature flags use. This leads to races while
setting up the state and removing it. Fix this by creating a
separate invalid repository for each test. The repository is created
through the API and broken afterwards to ensure its metadata records
are created by Praefect properly.

Remove the unnecessary test helper that is a leftover from the time all
tests shared a global storage.