Allow unauthenticated access to health checking service
Gitaly and Praefect are currently requiring authentication for
accessing health checking service. Health checks are not a sensitive
feature and it would be convenient for users to be able to health
check Gitaly without being authenticated. This commit disables
authentication for health checking service by wrapping with a type
that overrides the authentication check with a no-op one.

Changelog: fixed