Allow for graceful shutdowns
Previously, `graceful_stop_timeout` was only respected when a SIGHUP is
signaled to Gitaly, other signals such as SIGTERM or SIGINT caused the
process to shutdown immediately, interrupting in-flight requests.

With this change, `graceful_stop_timeout` is respected regardless of how
the process was signaled, which would also come in handy for when Gitaly is
run in K8s (due to how K8s manages pods by shutting them down instead of
reloading processes).

Related to https://gitlab.com/gitlab-org/gitaly/-/issues/4747