catfile: Remove `requestQueue.isBatchCommand`
When we introduced `ObjectReader` (which uses 'git cat-file
--batch-command'), we share the `requestQueue` implementation with
`ObjectInfoReader` (which uses 'git cat-file --batch-check') and
`ObjectContentReader` (which uses 'git cat-file --batch'). Now, both
`ObjectInfoReader` and `ObjectContentReader` are deprecated, so we can
strip away the itegration glue between the three.

The `isBatchCommand` was used to distinguish if the reader used was
`ObjectReader`, since this is now the only situation, we can remove the
fields and associated code around it.