git: Wire up Git v2.44 execution environment
Wire up the Git v2.44 execution environment so that we can toggle
between the v2.43 and v2.44 release tracks. Major new features that
we're interested in include:

    - new `git replay` command

    - work on the new "reftable" ref backend

    - fixes to unblock WAL
      (see https://gitlab.com/gitlab-com/gitlab-OKRs/-/work_items/5708)

Some of the fixes to unblock WAL, have not made it into the official
v2.44.0 release, but were merged to master afterwards and have been
added to our gitlab-git-v2.44 branch and tagged with the v2.44.0.gl1
tag that is used to build our v2.44 Git version.

There are naturally various other new features, bugfixes and performance
improvements.

The new Git v2.44 binary is protected by a feature flag.

Changelog: added