Extract housekeeping metrics out of housekeeping manager
Recently, the housekeeping manager is responsible for keeping track of
metrics of housekeeping tasks. In some later commits, we need to add
metric support to the WAL transaction manager. It will use a subset of
those metrics. Unfortunately, WAL transaction manager could not access
them directly due to circular dependency issue. In addition, it doesn't
make sense for the WAL to access internal metrics of housekeeping
manager.

This commit extracts housekeeping-related metrics to the generic
"housekeeping" package instead of "housekeeping/manager". A struct is
introduced to wrap those metrics. It will be re-used by both
housekeeping manager and WAL transaction manager.