Add housekeeping metrics to transaction manager
This MR lets the transaction manager record housekeeping metrics.
Unfortunately, there might be hundreds of thousands of transaction
managers, corresponding to the number of repositories (and their pool
members) in a node. Those metrics must be gathered by per-process
Prometheus collector. As a result, the partition manager is responsible
for holding metric recorders. It creates a wrapper object and passes
its pointer to the children's transaction managers. This approach is
flexible enough to add more metrics recording the activity of
transaction managers, not just housekeeping ones.

Asserting recorded metrics is a challenge. Although the Prometheus
library provides some test helpers to assert the collected metrics,
they are not flexible enough. As a result, this MR implements a
custom metric asserter for the transaction manager.