Offload test coverage report to CI artifact
Recently, each test job on CI reports the test coverage after the test
finishes. The list of per-function coverage is massive. It could be
up to 8k lines. Although the code coverage logs are collapsed by
default, Gitlab platform ignores that notation when the total line of
logs is too big. As a result, it's usually hard to look for failed
tests.

That list of test coverage is arguably useful, given we do report
per-package coverage. This commit offloads those logs to job artifacts
instead of dumping them directly to stdout.