Test alternate linking more realistically
TransactionManager's tests are currently staging alternate updates
directly in the transaction. This is done regardless whether the
alternate is actually in the transaction's snapshot. This is not
realistic as the object pool RPCs that would link a repository to
an alternate would always have the pool repository in the transaction's
snapshot. If not, they'd fail as they verify the alternate repository
actually exists.

This commit updates TransactionManager's tests to be more realistic by
ensuring that the alternate being linked is actually in the snapshot.
This way were testing closer to the actual scenarios we have. This
becomes more imporant as we're about to change the packing strategy
to only pack objects from the quarantine in the transaction's packfile.